test-alsa-midi: increase timeout
snd_seq_event_input() blocks until the event timestamp is reached,
according to the system clock. The ble_midi_decode() function
limits the time drift from system clock to 500ms. So in the worst
case, with 9 events to process, a total of 4500ms will pass just
blocked in calls to snd_seq_event_input(). Added to the total of
2100ms in usleep() calls, this gives a worst-case test time of
6600ms. The default default test timeout in libcheck is 4000ms,
and so we get occasional test timeout failures.

This commit increases the test-ble-midi timeout to 7000ms to avoid
these failures.