Reduce risk of transport release and acquisition colliding.
If a client re-opens a pcm immediately after closing the file
descriptor then there is a race between setting the "stopping"
flag and starting the new acquisition procedure. It is not
possible to make closing the fd and setting the flag into an
atomic operation, so this race in unavoidable and there is always
a non-zero risk that acquistion will collide with an in-progress
release procedure. This has been one of the causes of occasional
CI test failures in test-alsa-pcm:test-test_playback_hw_set_free.

The risk is significantly reduced by applying a short delay into
the acquisition procedure if the transport is not already idle.